Ajustes de APIs

+++ src/main/java/br/com/fiap/soat/grupo48/infrastructure/adapter/driver/rest/ProdutoController.java
@@ -2,13 +2,16 @@ package br.com.fiap.soat.grupo48.infrastructure.adapter.driver.rest;
 
 import br.com.fiap.soat.grupo48.application.produto.model.Produto;
 import br.com.fiap.soat.grupo48.application.produto.port.api.ProdutoServicePort;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
 @RestController
-@RequestMapping("produtos")
+@RequestMapping("api/produtos")
 public class ProdutoController {
     private final ProdutoServicePort produtoServicePort;
 
@@ -17,27 +20,40 @@ public class ProdutoController {
     }
 
     @GetMapping
-    List<Produto> getProdutos() {
-        return  this.produtoServicePort.buscarProdutos();
+    public ResponseEntity<List<Produto>> getProdutos() {
+        List<Produto> produtos = this.produtoServicePort.buscarProdutos();
+        return new ResponseEntity<>(produtos, HttpStatus.OK);
     }
 
     @GetMapping(value = "/{codigo}")
-    Produto getProduto(@PathVariable UUID codigo) {
-        return this.produtoServicePort.buscarPeloCodigo(codigo);
+    public ResponseEntity<Produto> getProduto(@PathVariable UUID codigo) {
+        Produto produto = this.produtoServicePort.buscarPeloCodigo(codigo);
+        if (Objects.isNull(produto)) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(produto, HttpStatus.OK);
     }
 
     @PostMapping
-    void criarProduto(@RequestBody Produto produto) {
-        this.produtoServicePort.criarProduto(produto);
+    public ResponseEntity<Produto> criarProduto(@RequestBody Produto produto) {
+        Produto produtoSave = this.produtoServicePort.criarProduto(produto);
+        return new ResponseEntity<>(produtoSave, HttpStatus.CREATED);
     }
     @PutMapping(value = "/{codigo}")
-    void atualizarProduto(@PathVariable UUID codigo, @RequestBody Produto produto) {
-        this.produtoServicePort.atualizarProduto(codigo,produto);
+    public ResponseEntity<Produto> atualizarProduto(@PathVariable UUID codigo, @RequestBody Produto produto) {
+        if (Objects.isNull(this.produtoServicePort.buscarPeloCodigo(codigo))) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        Produto produtoAtualizado = this.produtoServicePort.atualizarProduto(codigo,produto);
+        return new ResponseEntity<>(produtoAtualizado, HttpStatus.OK);
     }
 
     @DeleteMapping(value = "{codigo}")
-    void excluirProduto(@PathVariable UUID codigo) {
+    public ResponseEntity<Void> excluirProduto(@PathVariable UUID codigo) {
+        if (Objects.isNull(this.produtoServicePort.buscarPeloCodigo(codigo))) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         this.produtoServicePort.excluirProduto(codigo);
-
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
